+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-27T12:52:06.043Z")
+    created: Optional[datetime] = parse_datetime("2025-05-28T06:21:42.680Z")
     r"""The date and time the resource was created"""