+++ packages/orq-rc/RELEASES.md
@@ -1549,3 +1549,13 @@ Based on:
 - [python v3.6.0-rc.40] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.40] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.40 - packages/orq-rc
+
+## 2025-05-28 06:23:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.552.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.6.0-rc.41] packages/orq-rc
+### Releases
+- [PyPI v3.6.0-rc.41] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.41 - packages/orq-rc